requirements.txt
chromadb==0.6.3
huggingface_hub==0.31.1
langchain==0.3.25
langchain_chroma==0.2.3
langchain_community==0.3.24
langchain_core==0.3.59
langchain_huggingface==0.2.0
langchain_openai==0.3.16
loguru==0.7.3
Pillow==11.2.1
pysqlite_binary==0.5.1.3380300
python-dotenv==1.1.0
streamlit==1.45.0
streamlit_pdf_viewer==0.0.23
tiktoken==0.9.0
torch~=2.2.2
langsmith~=0.3.39
outlines~=0.2.3
pydantic~=2.11.4

pandas~=2.2.3
transformers~=4.51.3
openai~=1.77.0

openevals~=0.1.0

nltk~=3.9.1

kiwipiepy~=0.21.0
